fix rotate_gradients
