+++ qsiprep/workflows/dwi/hmc.py
@@ -473,8 +473,7 @@ def init_hmc_model_iteration_wf(modelname, transform, precision="coarse", name="
         (predict_dwis, outputnode, [('predicted_image', 'predicted_dwis')]),
         (post_bvec_transforms, outputnode, [('bvecs', 'aligned_bvecs')]),
         (register_to_predicted, outputnode, [('warped_image', 'aligned_dwis'),
-                                             (('forward_transforms', _list_squeeze),
-                                              'hmc_transforms')])
+                                             ('forward_transforms', 'hmc_transforms')])
     ])
 
     return workflow
@@ -685,19 +684,4 @@ def init_dwi_model_hmc_wf(modelname, transform, mem_gb, omp_nthreads,
 
 
 def _list_squeeze(in_list):
-    from typing import Iterable
-    from pathlib import Path
-    def flatten(items):
-        """Yield items from any nested iterable; see
-        Beazley, D. and B. Jones. Recipe 4.14, Python Cookbook 3rd Ed.,
-        O'Reilly Media Inc. Sebastopol, CA: 2013..
-
-        https://stackoverflow.com/questions/952914/how-do-i-make-a-flat-list-out-of-a-list-of-lists
-        """
-        for x in items:
-            if isinstance(x, Iterable) and not isinstance(x, (str, bytes, Path)):
-                for sub_x in flatten(x):
-                    yield sub_x
-            else:
-                yield x
-    return list(flatten(in_list))
+    return [item[0] for item in in_list]
